requirements.txt
Flask==3.0.0
mercadopago==2.2.0
fpdf==1.7.2
python-dotenv==1.0.0
Werkzeug==3.0.1
